+++ packages/gatsby-mdx/utils/create-mdx-node.js
@@ -1,8 +1,8 @@
 const crypto = require("crypto");
 const merge = require("lodash/merge");
-const defaultOptions = require("./utils/default-options");
-const mdx = require("./utils/mdx");
-const extractExports = require("./utils/extract-exports");
+const defaultOptions = require("./default-options");
+const mdx = require("./mdx");
+const extractExports = require("./extract-exports");
 
 module.exports = async (
   { node, transform, getNode, createNodeId },
